rename deepk to dlkoopman
