+++ README.md
@@ -42,7 +42,7 @@ The Deep Koopman system in this package performs three tasks: (TODO rework)
 - *Linearity* (`lin`): Learn a Koopman matrix which can operate on the initial encoded state $y_0$ to yield approximations $\{y_1',y_2',\cdots\}$ to the actual values $\{y_1,y_2,\cdots\}$, as well as predict unknown $y_t'$ for values of $t$ not in the given data.
 - *Prediction* (`pred`): $\{y_1',y_2',\cdots\}$ are decoded to predict approximations $\{\hat{x}_1',\hat{x}_2',\cdots\}$ to the actual values $\{x_1,x_2,\cdots\}$, as well as predict unknown $\hat{x}_t'$ for values of $t$ not in the given data. This is the task we care about the most.
 <figure><center>
-<img src="deepk_system.png" width=750/>
+<img src="dlkoopman_system.png" width=750/>
 </center></figure>
