+++ dlkoopman/utils.py
