+++ docs/config.html
@@ -4,7 +4,7 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
 <meta name="generator" content="pdoc 0.10.0" />
-<title>deepk.config API documentation</title>
+<title>dlkoopman.config API documentation</title>
 <meta name="description" content="Configuration options" />
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/sanitize.min.css" integrity="sha256-PK9q560IAAa6WVRRh76LtCaI8pjTJ2z11v0miyNNjrs=" crossorigin>
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/typography.min.css" integrity="sha256-7l/o7C8jubJiy74VsKTidCy1yBkRtiUGbVkYBylBqUg=" crossorigin>
@@ -20,7 +20,7 @@
 <main>
 <article id="content">
 <header>
-<h1 class="title">Module <code>deepk.config</code></h1>
+<h1 class="title">Module <code>dlkoopman.config</code></h1>
 </header>
 <section id="section-intro">
 <p>Configuration options</p>
@@ -30,7 +30,7 @@
 <section>
 <h2 class="section-title" id="header-variables">Global variables</h2>
 <dl>
-<dt id="deepk.config.precision"><code class="name">var <span class="ident">precision</span></code></dt>
+<dt id="dlkoopman.config.precision"><code class="name">var <span class="ident">precision</span></code></dt>
 <dd>
 <div class="desc"><p>Numerical precision of tensors.</p>
 <p><strong>Options</strong> (in increasing order of precision): <code>"half"</code> / <code>"float"</code> / <code>"double"</code>.</p>
@@ -38,13 +38,13 @@
 <h2 id="note">Note</h2>
 <p>Setting <code>precision = "double"</code> may help slightly with numerical instabilities, however, may also lead to inefficient GPU runtimes.</p></div>
 </dd>
-<dt id="deepk.config.use_cuda"><code class="name">var <span class="ident">use_cuda</span></code></dt>
+<dt id="dlkoopman.config.use_cuda"><code class="name">var <span class="ident">use_cuda</span></code></dt>
 <dd>
 <div class="desc"><p>If <code>True</code>, tensor computations when take place on CuDA GPUs whenever available.</p>
 <p><strong>Options</strong>: <code>True</code> / <code>False</code></p>
 <p><strong>Default</strong>: <code>True</code></p></div>
 </dd>
-<dt id="deepk.config.normalize_Xdata"><code class="name">var <span class="ident">normalize_Xdata</span></code></dt>
+<dt id="dlkoopman.config.normalize_Xdata"><code class="name">var <span class="ident">normalize_Xdata</span></code></dt>
 <dd>
 <div class="desc"><p>If <code>True</code>, all input states (training, validation, test) are divided by the maximum absolute value in the training data.</p>
 <p><strong>Options</strong>: <code>True</code> / <code>False</code></p>
@@ -58,7 +58,7 @@ This results in the singular and eigen vectors remaining the same.</p>
 <h2 id="caution">Caution</h2>
 <p>Setting this to <code>normalize_Xdata=False</code> may end the run by leading to imaginary parts of tensors reaching values where the loss function depends on the phase (in <code>torch &gt;= 1.11</code>, this leads to <code>"RuntimeError: linalg_eig_backward: The eigenvectors in the complex case are specified up to multiplication by e^{i phi}. The specified loss function depends on this quantity, so it is ill-defined"</code>). The only benefit to setting this to <code>False</code> is that if the run successfully completes, the final error metrics such as ANAE are slightly more accurate since they are reported on the un-normalized data values.</p></div>
 </dd>
-<dt id="deepk.config.use_exact_eigenvectors"><code class="name">var <span class="ident">use_exact_eigenvectors</span></code></dt>
+<dt id="dlkoopman.config.use_exact_eigenvectors"><code class="name">var <span class="ident">use_exact_eigenvectors</span></code></dt>
 <dd>
 <div class="desc"><p>If <code>True</code>, the exact eigenvectors of the Koopman matrix are used in <code>StatePredictor</code>, if <code>False</code>, the projected eigenvectors are used.</p>
 <p><strong>Options</strong>: <code>True</code> / <code>False</code></p>
@@ -66,7 +66,7 @@ This results in the singular and eigen vectors remaining the same.</p>
 <h2 id="notes">Notes</h2>
 <p>For a discussion on exact and projected eigenvectors, see <a href="https://arxiv.org/abs/1312.0041">Tu et al</a> or <a href="https://epubs.siam.org/doi/pdf/10.1137/1.9781611974508.ch1">Chapter 1 of Kutz et al</a>. The basic idea is that using exact eigenvectors is more accurate, but their computation may become less numerically stable than that of projected eigenvectors.</p></div>
 </dd>
-<dt id="deepk.config.sigma_threshold"><code class="name">var <span class="ident">sigma_threshold</span></code></dt>
+<dt id="dlkoopman.config.sigma_threshold"><code class="name">var <span class="ident">sigma_threshold</span></code></dt>
 <dd>
 <div class="desc"><p>When computing the SVD in <code>StatePredictor</code>, singular values lower than this will be reported, since they can be a possible cause of unstable gradients.</p>
 <p><strong>Options</strong>: Any numerical value.</p>
@@ -87,16 +87,16 @@ This results in the singular and eigen vectors remaining the same.</p>
 <ul id="index">
 <li><h3>Super-module</h3>
 <ul>
-<li><code><a title="deepk" href="index.html">deepk</a></code></li>
+<li><code><a title="dlkoopman" href="index.html">dlkoopman</a></code></li>
 </ul>
 </li>
 <li><h3><a href="#header-variables">Global variables</a></h3>
 <ul class="">
-<li><code><a title="deepk.config.precision" href="#deepk.config.precision">precision</a></code></li>
-<li><code><a title="deepk.config.use_cuda" href="#deepk.config.use_cuda">use_cuda</a></code></li>
-<li><code><a title="deepk.config.normalize_Xdata" href="#deepk.config.normalize_Xdata">normalize_Xdata</a></code></li>
-<li><code><a title="deepk.config.use_exact_eigenvectors" href="#deepk.config.use_exact_eigenvectors">use_exact_eigenvectors</a></code></li>
-<li><code><a title="deepk.config.sigma_threshold" href="#deepk.config.sigma_threshold">sigma_threshold</a></code></li>
+<li><code><a title="dlkoopman.config.precision" href="#dlkoopman.config.precision">precision</a></code></li>
+<li><code><a title="dlkoopman.config.use_cuda" href="#dlkoopman.config.use_cuda">use_cuda</a></code></li>
+<li><code><a title="dlkoopman.config.normalize_Xdata" href="#dlkoopman.config.normalize_Xdata">normalize_Xdata</a></code></li>
+<li><code><a title="dlkoopman.config.use_exact_eigenvectors" href="#dlkoopman.config.use_exact_eigenvectors">use_exact_eigenvectors</a></code></li>
+<li><code><a title="dlkoopman.config.sigma_threshold" href="#dlkoopman.config.sigma_threshold">sigma_threshold</a></code></li>
 </ul>
 </li>
 </ul>
