+++ docs/hyp_search.html
@@ -4,7 +4,7 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
 <meta name="generator" content="pdoc 0.10.0" />
-<title>deepk.hyp_search API documentation</title>
+<title>dlkoopman.hyp_search API documentation</title>
 <meta name="description" content="Run hyperparameter search." />
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/sanitize.min.css" integrity="sha256-PK9q560IAAa6WVRRh76LtCaI8pjTJ2z11v0miyNNjrs=" crossorigin>
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/typography.min.css" integrity="sha256-7l/o7C8jubJiy74VsKTidCy1yBkRtiUGbVkYBylBqUg=" crossorigin>
@@ -20,7 +20,7 @@
 <main>
 <article id="content">
 <header>
-<h1 class="title">Module <code>deepk.hyp_search</code></h1>
+<h1 class="title">Module <code>dlkoopman.hyp_search</code></h1>
 </header>
 <section id="section-intro">
 <p>Run hyperparameter search.</p>
@@ -32,7 +32,7 @@
 <section>
 <h2 class="section-title" id="header-functions">Functions</h2>
 <dl>
-<dt id="deepk.hyp_search.run_hyp_search"><code class="name flex">
+<dt id="dlkoopman.hyp_search.run_hyp_search"><code class="name flex">
 <span>def <span class="ident">run_hyp_search</span></span>(<span>dh, hyp_options, numruns=None, avg_ignore_initial_epochs=0, sort_key='avg_pred_anae_va') ‑> pathlib.Path</span>
 </code></dt>
 <dd>
@@ -117,12 +117,12 @@
 <ul id="index">
 <li><h3>Super-module</h3>
 <ul>
-<li><code><a title="deepk" href="index.html">deepk</a></code></li>
+<li><code><a title="dlkoopman" href="index.html">dlkoopman</a></code></li>
 </ul>
 </li>
 <li><h3><a href="#header-functions">Functions</a></h3>
 <ul class="">
-<li><code><a title="deepk.hyp_search.run_hyp_search" href="#deepk.hyp_search.run_hyp_search">run_hyp_search</a></code></li>
+<li><code><a title="dlkoopman.hyp_search.run_hyp_search" href="#dlkoopman.hyp_search.run_hyp_search">run_hyp_search</a></code></li>
 </ul>
 </li>
 </ul>
