+++ docs/index.html
@@ -4,7 +4,7 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
 <meta name="generator" content="pdoc 0.10.0" />
-<title>deepk API documentation</title>
+<title>dlkoopman API documentation</title>
 <meta name="description" content="" />
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/sanitize.min.css" integrity="sha256-PK9q560IAAa6WVRRh76LtCaI8pjTJ2z11v0miyNNjrs=" crossorigin>
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/typography.min.css" integrity="sha256-7l/o7C8jubJiy74VsKTidCy1yBkRtiUGbVkYBylBqUg=" crossorigin>
@@ -20,38 +20,38 @@
 <main>
 <article id="content">
 <header>
-<h1 class="title">Package <code>deepk</code></h1>
+<h1 class="title">Package <code>dlkoopman</code></h1>
 </header>
 <section id="section-intro">
 </section>
 <section>
 <h2 class="section-title" id="header-submodules">Sub-modules</h2>
 <dl>
-<dt><code class="name"><a title="deepk.config" href="config.html">deepk.config</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.config" href="config.html">dlkoopman.config</a></code></dt>
 <dd>
 <div class="desc"><p>Configuration options</p></div>
 </dd>
-<dt><code class="name"><a title="deepk.hyp_search" href="hyp_search.html">deepk.hyp_search</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.hyp_search" href="hyp_search.html">dlkoopman.hyp_search</a></code></dt>
 <dd>
 <div class="desc"><p>Run hyperparameter search.</p></div>
 </dd>
-<dt><code class="name"><a title="deepk.metrics" href="metrics.html">deepk.metrics</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.metrics" href="metrics.html">dlkoopman.metrics</a></code></dt>
 <dd>
 <div class="desc"><p>Loss and error functions, used to optimize models and report their performance.</p></div>
 </dd>
-<dt><code class="name"><a title="deepk.nets" href="nets.html">deepk.nets</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.nets" href="nets.html">dlkoopman.nets</a></code></dt>
 <dd>
 <div class="desc"><p>Neural nets used inside the models.</p></div>
 </dd>
-<dt><code class="name"><a title="deepk.state_predictor" href="state_predictor.html">deepk.state_predictor</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.state_predictor" href="state_predictor.html">dlkoopman.state_predictor</a></code></dt>
 <dd>
 <div class="desc"><p><strong>StatePredictor class, including its DataHandler</strong></p></div>
 </dd>
-<dt><code class="name"><a title="deepk.trajectory_predictor" href="trajectory_predictor.html">deepk.trajectory_predictor</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.trajectory_predictor" href="trajectory_predictor.html">dlkoopman.trajectory_predictor</a></code></dt>
 <dd>
 <div class="desc"><p><strong>TrajectoryPredictor class, including its DataHandler</strong></p></div>
 </dd>
-<dt><code class="name"><a title="deepk.utils" href="utils.html">deepk.utils</a></code></dt>
+<dt><code class="name"><a title="dlkoopman.utils" href="utils.html">dlkoopman.utils</a></code></dt>
 <dd>
 <div class="desc"><p>Utilities</p></div>
 </dd>
@@ -72,13 +72,13 @@
 <ul id="index">
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
-<li><code><a title="deepk.config" href="config.html">deepk.config</a></code></li>
-<li><code><a title="deepk.hyp_search" href="hyp_search.html">deepk.hyp_search</a></code></li>
-<li><code><a title="deepk.metrics" href="metrics.html">deepk.metrics</a></code></li>
-<li><code><a title="deepk.nets" href="nets.html">deepk.nets</a></code></li>
-<li><code><a title="deepk.state_predictor" href="state_predictor.html">deepk.state_predictor</a></code></li>
-<li><code><a title="deepk.trajectory_predictor" href="trajectory_predictor.html">deepk.trajectory_predictor</a></code></li>
-<li><code><a title="deepk.utils" href="utils.html">deepk.utils</a></code></li>
+<li><code><a title="dlkoopman.config" href="config.html">dlkoopman.config</a></code></li>
+<li><code><a title="dlkoopman.hyp_search" href="hyp_search.html">dlkoopman.hyp_search</a></code></li>
+<li><code><a title="dlkoopman.metrics" href="metrics.html">dlkoopman.metrics</a></code></li>
+<li><code><a title="dlkoopman.nets" href="nets.html">dlkoopman.nets</a></code></li>
+<li><code><a title="dlkoopman.state_predictor" href="state_predictor.html">dlkoopman.state_predictor</a></code></li>
+<li><code><a title="dlkoopman.trajectory_predictor" href="trajectory_predictor.html">dlkoopman.trajectory_predictor</a></code></li>
+<li><code><a title="dlkoopman.utils" href="utils.html">dlkoopman.utils</a></code></li>
 </ul>
 </li>
 </ul>
