+++ docs/metrics.html
@@ -4,7 +4,7 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
 <meta name="generator" content="pdoc 0.10.0" />
-<title>deepk.metrics API documentation</title>
+<title>dlkoopman.metrics API documentation</title>
 <meta name="description" content="Loss and error functions, used to optimize models and report their performance." />
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/sanitize.min.css" integrity="sha256-PK9q560IAAa6WVRRh76LtCaI8pjTJ2z11v0miyNNjrs=" crossorigin>
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/typography.min.css" integrity="sha256-7l/o7C8jubJiy74VsKTidCy1yBkRtiUGbVkYBylBqUg=" crossorigin>
@@ -20,7 +20,7 @@
 <main>
 <article id="content">
 <header>
-<h1 class="title">Module <code>deepk.metrics</code></h1>
+<h1 class="title">Module <code>dlkoopman.metrics</code></h1>
 </header>
 <section id="section-intro">
 <p>Loss and error functions, used to optimize models and report their performance.</p>
@@ -32,7 +32,7 @@
 <section>
 <h2 class="section-title" id="header-functions">Functions</h2>
 <dl>
-<dt id="deepk.metrics.anae"><code class="name flex">
+<dt id="dlkoopman.metrics.anae"><code class="name flex">
 <span>def <span class="ident">anae</span></span>(<span>ref, new) ‑> torch.Tensor</span>
 </code></dt>
 <dd>
@@ -54,7 +54,7 @@ Note that:</p>
 <h2 id="returns">Returns</h2>
 <p><strong>anae</strong> (<em>torch scalar</em>) - In percentage.</p></div>
 </dd>
-<dt id="deepk.metrics.overall_anae"><code class="name flex">
+<dt id="dlkoopman.metrics.overall_anae"><code class="name flex">
 <span>def <span class="ident">overall_anae</span></span>(<span>X, Y, Xr, Ypred, Xpred) ‑> dict[str, torch.Tensor]</span>
 </code></dt>
 <dd>
@@ -83,7 +83,7 @@ Note that:</p>
 - Key <strong>'lin'</strong>: (<em>torch scalar</em>) - Linearity error between <code>Y</code> and <code>Ypred</code>.
 - Key <strong>'pred'</strong>: (<em>torch scalar</em>) - Prediction error between <code>X</code> and <code>Xpred</code>.</p></div>
 </dd>
-<dt id="deepk.metrics.overall_loss"><code class="name flex">
+<dt id="dlkoopman.metrics.overall_loss"><code class="name flex">
 <span>def <span class="ident">overall_loss</span></span>(<span>X, Y, Xr, Ypred, Xpred, decoder_loss_weight) ‑> dict[str, torch.Tensor]</span>
 </code></dt>
 <dd>
@@ -129,14 +129,14 @@ Note that:</p>
 <ul id="index">
 <li><h3>Super-module</h3>
 <ul>
-<li><code><a title="deepk" href="index.html">deepk</a></code></li>
+<li><code><a title="dlkoopman" href="index.html">dlkoopman</a></code></li>
 </ul>
 </li>
 <li><h3><a href="#header-functions">Functions</a></h3>
 <ul class="">
-<li><code><a title="deepk.metrics.anae" href="#deepk.metrics.anae">anae</a></code></li>
-<li><code><a title="deepk.metrics.overall_anae" href="#deepk.metrics.overall_anae">overall_anae</a></code></li>
-<li><code><a title="deepk.metrics.overall_loss" href="#deepk.metrics.overall_loss">overall_loss</a></code></li>
+<li><code><a title="dlkoopman.metrics.anae" href="#dlkoopman.metrics.anae">anae</a></code></li>
+<li><code><a title="dlkoopman.metrics.overall_anae" href="#dlkoopman.metrics.overall_anae">overall_anae</a></code></li>
+<li><code><a title="dlkoopman.metrics.overall_loss" href="#dlkoopman.metrics.overall_loss">overall_loss</a></code></li>
 </ul>
 </li>
 </ul>
