+++ docs/nets.html
@@ -4,7 +4,7 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
 <meta name="generator" content="pdoc 0.10.0" />
-<title>deepk.nets API documentation</title>
+<title>dlkoopman.nets API documentation</title>
 <meta name="description" content="Neural nets used inside the models." />
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/sanitize.min.css" integrity="sha256-PK9q560IAAa6WVRRh76LtCaI8pjTJ2z11v0miyNNjrs=" crossorigin>
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/typography.min.css" integrity="sha256-7l/o7C8jubJiy74VsKTidCy1yBkRtiUGbVkYBylBqUg=" crossorigin>
@@ -20,7 +20,7 @@
 <main>
 <article id="content">
 <header>
-<h1 class="title">Module <code>deepk.nets</code></h1>
+<h1 class="title">Module <code>dlkoopman.nets</code></h1>
 </header>
 <section id="section-intro">
 <p>Neural nets used inside the models.</p>
@@ -34,7 +34,7 @@
 <section>
 <h2 class="section-title" id="header-classes">Classes</h2>
 <dl>
-<dt id="deepk.nets.AutoEncoder"><code class="flex name class">
+<dt id="dlkoopman.nets.AutoEncoder"><code class="flex name class">
 <span>class <span class="ident">AutoEncoder</span></span>
 <span>(</span><span>input_size, encoded_size, encoder_hidden_layers=[], decoder_hidden_layers=[], batch_norm=False)</span>
 </code></dt>
@@ -69,7 +69,7 @@
 </ul></div>
 <h3>Methods</h3>
 <dl>
-<dt id="deepk.nets.AutoEncoder.forward"><code class="name flex">
+<dt id="dlkoopman.nets.AutoEncoder.forward"><code class="name flex">
 <span>def <span class="ident">forward</span></span>(<span>self, X) ‑> tuple[torch.Tensor, torch.Tensor]</span>
 </code></dt>
 <dd>
@@ -90,7 +90,7 @@
 </dd>
 </dl>
 </dd>
-<dt id="deepk.nets.Knet"><code class="flex name class">
+<dt id="dlkoopman.nets.Knet"><code class="flex name class">
 <span>class <span class="ident">Knet</span></span>
 <span>(</span><span>size)</span>
 </code></dt>
@@ -107,7 +107,7 @@
 </ul></div>
 <h3>Methods</h3>
 <dl>
-<dt id="deepk.nets.Knet.forward"><code class="name flex">
+<dt id="dlkoopman.nets.Knet.forward"><code class="name flex">
 <span>def <span class="ident">forward</span></span>(<span>self, X) ‑> torch.Tensor</span>
 </code></dt>
 <dd>
@@ -134,21 +134,21 @@
 <ul id="index">
 <li><h3>Super-module</h3>
 <ul>
-<li><code><a title="deepk" href="index.html">deepk</a></code></li>
+<li><code><a title="dlkoopman" href="index.html">dlkoopman</a></code></li>
 </ul>
 </li>
 <li><h3><a href="#header-classes">Classes</a></h3>
 <ul>
 <li>
-<h4><code><a title="deepk.nets.AutoEncoder" href="#deepk.nets.AutoEncoder">AutoEncoder</a></code></h4>
+<h4><code><a title="dlkoopman.nets.AutoEncoder" href="#dlkoopman.nets.AutoEncoder">AutoEncoder</a></code></h4>
 <ul class="">
-<li><code><a title="deepk.nets.AutoEncoder.forward" href="#deepk.nets.AutoEncoder.forward">forward</a></code></li>
+<li><code><a title="dlkoopman.nets.AutoEncoder.forward" href="#dlkoopman.nets.AutoEncoder.forward">forward</a></code></li>
 </ul>
 </li>
 <li>
-<h4><code><a title="deepk.nets.Knet" href="#deepk.nets.Knet">Knet</a></code></h4>
+<h4><code><a title="dlkoopman.nets.Knet" href="#dlkoopman.nets.Knet">Knet</a></code></h4>
 <ul class="">
-<li><code><a title="deepk.nets.Knet.forward" href="#deepk.nets.Knet.forward">forward</a></code></li>
+<li><code><a title="dlkoopman.nets.Knet.forward" href="#dlkoopman.nets.Knet.forward">forward</a></code></li>
 </ul>
 </li>
 </ul>
