+++ docs/utils.html
@@ -4,7 +4,7 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
 <meta name="generator" content="pdoc 0.10.0" />
-<title>deepk.utils API documentation</title>
+<title>dlkoopman.utils API documentation</title>
 <meta name="description" content="Utilities" />
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/sanitize.min.css" integrity="sha256-PK9q560IAAa6WVRRh76LtCaI8pjTJ2z11v0miyNNjrs=" crossorigin>
 <link rel="preload stylesheet" as="style" href="https://cdnjs.cloudflare.com/ajax/libs/10up-sanitize.css/11.0.1/typography.min.css" integrity="sha256-7l/o7C8jubJiy74VsKTidCy1yBkRtiUGbVkYBylBqUg=" crossorigin>
@@ -20,7 +20,7 @@
 <main>
 <article id="content">
 <header>
-<h1 class="title">Module <code>deepk.utils</code></h1>
+<h1 class="title">Module <code>dlkoopman.utils</code></h1>
 </header>
 <section id="section-intro">
 <p>Utilities</p>
@@ -32,7 +32,7 @@
 <section>
 <h2 class="section-title" id="header-functions">Functions</h2>
 <dl>
-<dt id="deepk.utils.stable_svd"><code class="name flex">
+<dt id="dlkoopman.utils.stable_svd"><code class="name flex">
 <span>def <span class="ident">stable_svd</span></span>(<span>x) ‑> tuple[torch.Tensor, torch.Tensor, torch.Tensor]</span>
 </code></dt>
 <dd>
@@ -47,7 +47,7 @@
 <li><strong>V</strong> (<em>torch.Tensor, shape=(n, min(m,n))</em>)</li>
 </ul></div>
 </dd>
-<dt id="deepk.utils.plot_stats"><code class="name flex">
+<dt id="dlkoopman.utils.plot_stats"><code class="name flex">
 <span>def <span class="ident">plot_stats</span></span>(<span>model, perfs=['pred_anae'], start_epoch=1, fontsize=12)</span>
 </code></dt>
 <dd>
@@ -70,7 +70,7 @@
 <h2 id="effects">Effects</h2>
 <p>Creates plots for each <code>perf</code> and saves their png file(s) to <code>"./plot_&lt;model.uuid&gt;_&lt;perf&gt;.png"</code>.</p></div>
 </dd>
-<dt id="deepk.utils.set_seed"><code class="name flex">
+<dt id="dlkoopman.utils.set_seed"><code class="name flex">
 <span>def <span class="ident">set_seed</span></span>(<span>seed)</span>
 </code></dt>
 <dd>
@@ -93,14 +93,14 @@
 <ul id="index">
 <li><h3>Super-module</h3>
 <ul>
-<li><code><a title="deepk" href="index.html">deepk</a></code></li>
+<li><code><a title="dlkoopman" href="index.html">dlkoopman</a></code></li>
 </ul>
 </li>
 <li><h3><a href="#header-functions">Functions</a></h3>
 <ul class="">
-<li><code><a title="deepk.utils.stable_svd" href="#deepk.utils.stable_svd">stable_svd</a></code></li>
-<li><code><a title="deepk.utils.plot_stats" href="#deepk.utils.plot_stats">plot_stats</a></code></li>
-<li><code><a title="deepk.utils.set_seed" href="#deepk.utils.set_seed">set_seed</a></code></li>
+<li><code><a title="dlkoopman.utils.stable_svd" href="#dlkoopman.utils.stable_svd">stable_svd</a></code></li>
+<li><code><a title="dlkoopman.utils.plot_stats" href="#dlkoopman.utils.plot_stats">plot_stats</a></code></li>
+<li><code><a title="dlkoopman.utils.set_seed" href="#dlkoopman.utils.set_seed">set_seed</a></code></li>
 </ul>
 </li>
 </ul>
